chore: remove circular dependencies with button.ts
