+++ packages/plugin-core/src/test/suite-integ/NoteLookupCommand.test.ts
@@ -37,7 +37,6 @@ import {
 } from "../../commands/NoteLookupCommand";
 import {
   ButtonType,
-  DendronBtn,
   HorizontalSplitBtn,
   JournalBtn,
   ScratchBtn,
@@ -45,6 +44,7 @@ import {
   SelectionExtractBtn,
   TaskBtn,
 } from "../../components/lookup/buttons";
+import { DendronBtn } from "../../components/lookup/ButtonTypes";
 import { CREATE_NEW_LABEL } from "../../components/lookup/constants";
 import {
   DendronQuickPickerV2,
