+++ packages/plugin-core/src/views/LookupView.ts
@@ -12,8 +12,9 @@ import { WebViewUtils } from "./utils";
 import { LookupControllerV3 } from "../components/lookup/LookupControllerV3";
 import { PickerUtilsV2 } from "../components/lookup/utils";
 import { DendronQuickPickerV2 } from "../components/lookup/types";
-import { DendronBtn, getButtonCategory } from "../components/lookup/buttons";
+import { getButtonCategory } from "../components/lookup/buttons";
 import { getExtension } from "../workspace";
+import { DendronBtn } from "../components/lookup/ButtonTypes";
 
 export class LookupView implements vscode.WebviewViewProvider {
   public static readonly viewType = DendronTreeViewKey.LOOKUP_VIEW;
